+++ src/render/BasicMeshEffect.h
@@ -5,9 +5,7 @@
 
 #include "render/IRenderEffect.h"
 
-#include <CommonStates.h>
 #include <filesystem>
-#include <memory>
 #include <wrl/client.h>
 
 namespace lrender
@@ -26,7 +24,6 @@ class BasicMeshEffect final : public IRenderEffect
     [[nodiscard]] bool IsWireframe() const noexcept;
 
   private:
-    std::unique_ptr<DirectX::CommonStates> m_states;
     Microsoft::WRL::ComPtr<ID3D11VertexShader> m_vertexShader;
     Microsoft::WRL::ComPtr<ID3D11PixelShader> m_pixelShader;
     Microsoft::WRL::ComPtr<ID3D11InputLayout> m_inputLayout;