+++ src/render/EffectResource.h
@@ -13,6 +13,12 @@
 namespace lrender
 {
 
+enum class EffectResourceSizeMode
+{
+    MatchViewport,
+    Fixed
+};
+
 class EffectResource final
 {
   public:
@@ -24,7 +30,9 @@ class EffectResource final
     EffectResource& operator=(EffectResource&&) noexcept = default;
 
     void Resize(ID3D11Device* device, std::uint32_t width, std::uint32_t height);
+    void ResizeForViewport(ID3D11Device* device, std::uint32_t width, std::uint32_t height);
     void Reset() noexcept;
+    void SetSizeMode(EffectResourceSizeMode mode) noexcept;
     void BindAndClear(ID3D11DeviceContext* context, const float clearColor[4],
                       const EffectResource* additionalResource = nullptr) const;
 
@@ -33,10 +41,12 @@ class EffectResource final
     [[nodiscard]] ID3D11DepthStencilView* GetDepthStencilView() const noexcept;
     [[nodiscard]] std::uint32_t GetWidth() const noexcept;
     [[nodiscard]] std::uint32_t GetHeight() const noexcept;
+    [[nodiscard]] EffectResourceSizeMode GetSizeMode() const noexcept;
 
   private:
     std::uint32_t m_width = 0;
     std::uint32_t m_height = 0;
+    EffectResourceSizeMode m_sizeMode = EffectResourceSizeMode::MatchViewport;
     Microsoft::WRL::ComPtr<ID3D11Texture2D> m_colorTexture;
     Microsoft::WRL::ComPtr<ID3D11RenderTargetView> m_renderTargetView;
     Microsoft::WRL::ComPtr<ID3D11ShaderResourceView> m_shaderResourceView;