+++ src/render/SkyCubeEffect.cpp
@@ -3,6 +3,7 @@
  */
 
 #include "stdfx.h"
+#include "render/EffectManager.h"
 #include "render/SkyCubeEffect.h"
 #include <stdexcept>
 
@@ -26,7 +27,7 @@ void ThrowIfFailed(HRESULT result, const char* message)
 SkyCubeEffect::SkyCubeEffect(ID3D11Device* device, ID3D11DeviceContext* context,
                              const std::filesystem::path& shaderDirectory,
                              const std::filesystem::path& cubeMapPath)
-    : IRenderEffect(device, context), m_states(std::make_unique<DirectX::CommonStates>(Device()))
+    : IRenderEffect(device, context)
 {
     const auto vertexShader = LoadShader(shaderDirectory / L"SkyVS.cso");
     const auto pixelShader = LoadShader(shaderDirectory / L"SkyPS.cso");
@@ -36,7 +37,7 @@ SkyCubeEffect::SkyCubeEffect(ID3D11Device* device, ID3D11DeviceContext* context,
     ThrowIfFailed(Device()->CreatePixelShader(pixelShader->GetBufferPointer(), pixelShader->GetBufferSize(), nullptr,
                                               m_pixelShader.GetAddressOf()),
                   "Failed to create sky pixel shader");
-    m_cubeMapResource.LoadDDS(Device(), cubeMapPath);
+    m_cubeMapResource = EffectManager::Instance().CreateCubeMapResource(cubeMapPath);
 }
 
 void SkyCubeEffect::Bind(const EffectFrameContext& frame)
@@ -61,18 +62,19 @@ void SkyCubeEffect::Draw(const EffectFrameContext& frame)
     ID3D11DeviceContext* context = frame.DeviceContext();
     ID3D11ShaderResourceView* cubeMap = m_cubeMapResource.GetShaderResourceView();
     context->PSSetShaderResources(SkyTextureCubeSLOT, 1, &cubeMap);
-    ID3D11SamplerState* sampler = m_states->LinearClamp();
+    EffectManager& effectManager = EffectManager::Instance();
+    ID3D11SamplerState* sampler = effectManager.GetLinearClampSampler();
     context->PSSetSamplers(LinearClampSamplerSLOT, 1, &sampler);
     //在光栅化阶段关闭背面消隐（正面是立方体向外的面，但摄像机在内部）
-    context->RSSetState(m_states->CullNone());
+    effectManager.SetRasterizerMode(RasterizerMode::SolidCullNone);
     //在输出合并阶段设置深度/模板状态，深度比较函数为小于等于，以允许深度值为 1 的像素绘制
-    D3D11_DEPTH_STENCIL_DESC description{};
-    description.DepthEnable = TRUE;
-    description.DepthWriteMask = D3D11_DEPTH_WRITE_MASK_ZERO;
-    description.DepthFunc = D3D11_COMPARISON_LESS_EQUAL;
-    const auto depthStencilState = ViewManager::Instance().CreateDepthStencilState(description);
+	D3D11_DEPTH_STENCIL_DESC description{};
+	description.DepthEnable = TRUE;
+	description.DepthWriteMask = D3D11_DEPTH_WRITE_MASK_ZERO;
+	description.DepthFunc = D3D11_COMPARISON_LESS_EQUAL;
+    const auto depthStencilState = effectManager.CreateDepthStencilState(description);
+    context->OMSetDepthStencilState(depthStencilState.Get(),0);
     //绘制
-    context->OMSetDepthStencilState(depthStencilState.Get(), 0);
     context->Draw(3, 0);
     //还原
     ID3D11ShaderResourceView* nullResource = nullptr;